trim spaces and void blanks
we now trim redoundant blanks from templates if they come (out of <pre>, <script>, <textarea> tags).

it's not 100% of the unnecessray blanks.  but it still trims a lot.
I assume we gain at least 10% on the size of the file.

moreover, with today's commits, we gain about 0.02s per page creation (according to the times I read on the top),
and a simple "static" page takes ~0.15s to load.  so it's a 5% speedup, with commits that _simplify_ the code...
what do you people want more ?


git-archimport-id: opensource@polytechnique.org--2005/platal--mainline--0.9--patch-249