add a 4th option : tag_and_drop_spams, that discards mails that have a tag-spam greater than 0.999999
HTML RANT :

paragraphs are delimited with <p> ... </p>, not infamous <br />'s
<br /s> are used iff you want to break a line e.g. in a formular. but for the
sake of the readability of your page, *USE* <p>'s since they are adding vertical
spacing, adding a lot more of readability.


moreover, it is discouradged to use both <em> and <strong> on the same page for
the same meaning. it render the reading way more complicated. Just let the
skinner deal with that, you are the writer, you shouldn't do anything else than
organizing the DOM.


last point is the use of any heml element. those have names, and there is
plenty. It's stupid after all, with body/img/div/span/a/table + css you can do
everything you want. *but* there is more. why ? because it gives element a
semantics. SO DON'T USE <h3> JUST BECAUSE IT LOOKS LIKE <strong> IN YOUR SKIN.
this is an horror. h1..6 are *HEADINGS* and their semantics is to separate
topics and parts in the page.

the same apply to the previous discussion around <br />. br has to be used ONLY
AND ONLY IF you are *NOT* beggining a new paragraph, but INTENTIONNALY forcing
the break of the line. when you end a paragraph, html has a tag that has this
semantics : </p>. I don't want to see anymore <br />-delimited paragraphs.

thanks for reading and considering.



git-svn-id: svn+ssh://murphy/home/svn/platal/trunk@109 839d8a87-29fc-0310-9880-83ba4fa771e5