mostly code simplifications.
also: drop support for membres_only and advertise for events as it's not supported at all, easy to
re-add and just confuses users.

if we plan to add a feature, we have to add them fully or not, but not half, it's very confusing, and
lead to bloated forms for nothing.



git-svn-id: svn+ssh://murphy/home/svn/platal/trunk@523 839d8a87-29fc-0310-9880-83ba4fa771e5