tons of fixes and/or simplifications.
but that's really clear: current method to deal with pmwiki is *way* too
complicated, I can't understand a single line of that piece of code, and
how it interacts with the rest.

We don't need 10% of the complexity of what is currently developped.

moreover, users can import any smarty tags in the text, which is awful :|



git-svn-id: svn+ssh://murphy/home/svn/platal/trunk@548 839d8a87-29fc-0310-9880-83ba4fa771e5