Partially reverts 5594f3a9 ("jQuery.Autocomplete should be downloaded by the Makefile").
It turns out that the jquery.autocomplete.js downloaded on plugins.jquery.com is a different version, which is not compatible with the previous autocompleter.
Furthermore, the new version of our previous autocompleter is not backward compatible with non-fixed width autocompletion (which is problematic since we have autocompletion on /really/ long company names); since we'll have anyway to patch our autocomplete.js, we might as well keep the old version.

Signed-off-by: Vincent Zanotti <vincent.zanotti@polytechnique.org>