Remodels the registration process (part 2).
Part 1 consisted in the addition of reminder "boxes" at top of the /event page, which would propose the user several features (subscription to NL/AXLetter, ...) that were normally offered at the end of the registration process.
Part 2 consisted of the removal of the "end" of the registration process: password is now set during the initial part of the registration, and is re-checked after the user has clicked on the link in the registration email.

Signed-off-by: Vincent Zanotti <vincent.zanotti@m4x.org>