Add script for filling directory_name for xnet accounts (Closes #1178)
This script tries to guess the directory_name from the full_name.
Any part matching "(...)" or " - ..." will be considered as a comment,
    and added to the 'comment' field of the group this user is a member
    of (only if there is only one such group and no current comment).

All errors are displayed at the end of the run, the 'directory_name'
won't be fixed for them (still 'NULL'); this list can then be accessed
by simply running the script again.

Signed-off-by: Raphaël Barrois <raphael.barrois@polytechnique.org>