AX enabled version of the first page of the profile edition.
If the profile editor does not have directory_private privileges, hide
private informations.

Note: use <input type="hidden" /> to hide some informations without
breaking the code of the backend. I know this does not hide anything
to somebody who knows html, but remember: this is for the secretaries
of the AX... We can certainly do better in the future, but this requires
some hacking in the php code for phones, networking, addresses, jobs and
I'm too lazy for it.

Signed-off-by: Florent Bruneau <florent.bruneau@polytechnique.org>