Fix and improves the 'become user' feature.
You can now select an account type among all available types and see the
site with the permissions of this type. Leaving a SUID session go back to
the page where the SUID session has been started.

Note: The site seen by someone with this account type might be a bit
different since some parameters are independent from the type (like the
fact your account is attached to a profile).

Signed-off-by: Florent Bruneau <florent.bruneau@polytechnique.org>