New script to perform the upgrade of the site.
Instead of calling:
 % cd upgrade/{version}/
 % ./update.sh
The migration of the db can be done through:
 % upgrade/update.sh

The script will automatically discover the version of the upgrade scripts
to be run.

Note: my goal is to build a script that can be used to perform most of the
steps of the migration, not only the upgrade of the db and, if possible,
that can detect the needed upgrades (let say "what must I do to upgrade
this site that ran platal 0.10.1 to platal 1.0.1"). Feel free to
improve this script.

Signed-off-by: Florent Bruneau <florent.bruneau@polytechnique.org>