Fix download URL for openid.
Note: PHP-OpenId 2.1.3 is not available anymore. We need to test openid
authentication using the new version of openid.

Note2: Packaging from github suxx: it contains both a version number (in
the URL) and a commit-id (in the package content). This is just
unmaintainable and will break again and again in the future.

Signed-off-by: Florent Bruneau <florent.bruneau@polytechnique.org>