Fix PHP Notice: Array to string conversion in validate.inc.php
Validate::iterate() builds a SQL query that involves an array of WHERE filters.
When this array is empty, it was incorrectly kept as is, thus triggering a PHP
Notice.

Note: this error is not fatal as the following statement is a valid SQL query:
SELECT data, DATE_FORMAT(stamp, "%Y%m%d%H%i%s") FROM requests Array ORDER BY stamp;