Rename ax_mail flag to dn_best_mail for profile addresses
The name "ax_mail" is ambiguous with the visibility level "ax". In fact this
is only used to speed up the advanced search when an administrator exports a
list of addresses to prepare a mailing.

Therefore this denormalized flag is renamed "dn_best_mail", which is the best
known address to send mail.

This commit adds a comment in classes/address.php to describe how this best
address is chosen.