Add a cron which checks the best mail addresses which is stored
The cron script computes the selected mail address for each profile and warns
once a week if the result doesn't match the database data.

Also make Address::updateBestMail static public so that there is no need to
create an object per profile in the cron script and there is no code
duplication to find which address would be selected.