Allow adding accounts for old promotions in account admin page
When an alumni reports a typo in the name which was given for the
registration process (for example "Oliver" instead of "Olivier"), the
easiest solution consists in creating a new account.  The standard
interface couldn't be used for this as it doesn't allow promotions
before 1996 for X alumni.

Weaken this requirement to allow specifying a xorg ID instead of a
school ID.