Fix spamiful education comparison when recording profile changes
When modifying the profile of a person having an education without a
graduation year, a "warning" field is added to the submitted data which
is compared with the initial data.  Therefore a change of education on
the profile is recorded even though it is not the case.  Fix this by
adding this "warning" field to the initial data too.