Rework next_birthday computation
There were issues with the incremental method, as people born on a
February 29th got their birthday on the 28th every year.

Fix this by computing next_birthday from the absolute birth date.

Moreover when birthdate is modified on the profile, next_birthday needs
to be recomputed.  Detect such a case.