Redirect users of "send email" page to membres@ mailing lists
At the last X-Internet meeting, it has been reported that the dissuasive
text introduced back in March 2014 (commit ad47a0aa4c6a, "Add a
dissuasive message in the "send a mail to a group" page") was not clear
enough because "newsletter" does not sound like "sending an email to
group members".

Therefore add a notice about using membres@ mailing list instead, and
the newsletter for users really fond of the "<cher> <prenom>" feature.

While at it, make the checkbox option more dissuasive and understandable
in the group admin page.