Kill the possibility to send an email to all members of a group
This feature was buggy for large groups (the page was loading until
every email got sent) and confusing for some people who did not
understand that mailing lists allow using a mail client to write emails.

Because of this, kill the "send to all members" feature, to make people
use real mailing lists or the group newsletter.

If many users complain because of this commit, feel free to revert it
after understanding what they are missing.