mailinglist: Add 'sudo' mode.
This allows a system account (e.g cron) to connect on behalf of a user.
The MailingList operations are still performed with the permissions of
that user; it only allows bypassing the authentication.

Rationale: some users have no defined password (if it was compromised,
for instance)

Technically, the system will login as
<su_login>+<real_login>:<su_password>