Don't notify end users deathdate update.
The code tried to account for this; however, at some point in time, the
content of the 'field' column in 'profile_modification' was updated to
hold the field's display name instead of its code.
This change prevented the actual security code to work properly.

This commit will change the database storage to use a field's codename
instead of its display name.