Use the right French word to describe password hashing
In French, "crypter" can mean to put into a crypt, which may be found in
some churches. The right word would be for example "chiffrer" for
messages and "hacher" or "obtenir un condensat/une empreinte" for
passwords.

Write messages which do not use "crypter" and which non-technical people
can still understand.