Fix email address in profile modification email
When an administrator updates the email which is displayed in the
directory to an address which is neither a main address nor a
redirection one, the notification email which is currently sent tells
that the address has been modified to "new@example.org". This is the
value of option "New email address" in the profile modification page,
not the real value of the address which got added to the database.

Fix this by modifying the value of field "email_directory" in the
profile form for general information when field "email_directory_new"
contains a valid email address.