Move the password change page to auth.polytechnique.org
Some users continue to use it in order to change their passwords, even
though there is no chance for it to work. As the SMTP password is not
yet handled by auth.polytechnique.org, keep the page with a link to the
SMTP password change page, but add a link to auth.polytechnique.org.