Only show recent AX letters
There are too many AX letters for the admin pages and it was triggering
a 500 error (because of exhausted memory). The current implementation of
newsletter issues listing does not scale at all, but instead of
refactoring it in order to reduce the number of SQL statements,
introduce a reasonable time limit which speeds up a lot the AX pages.